pyproject.toml
[tool.poetry]
name = "alain_trans"
version = "0.1.0"
description = ""
authors = ["pierre kasongo <pierrekasongo88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.0"
logger = "^1.4"
DateTime = "^4.7"
pathlib = "^1.0.1"
alembic = "^1.8.1"
databases = "^0.6.1"
asyncpg = "^0.26.0"
psycopg2 = "^2.9.4"
PyJWT = "^2.6.0"
python-decouple = "^3.6"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
